Use skipWhenMissingDependencies for GRAD Trenches compatiblity
