--- addons/compat_sog/addon.toml
@@ -1,1 +1,0 @@
-preprocess = false
