+++ addons/compat_sog/subConfig/grad_trenches/script_component.hpp
@@ -1,0 +1,5 @@
+#include "\z\ace\addons\compat_sog\script_component.hpp"
+
+#define SUBCOMPONENT grad_trenches
+#undef ADDON
+#define ADDON TRIPLES(PREFIX,COMPONENT,SUBCOMPONENT)
